+++ MyMangaList/database/migrations/2021_06_14_132400_create_MangaLectureFini_table.php
@@ -16,7 +16,7 @@ class CreateMangaLectureFiniTable extends Migration
         Schema::create('MangaLectureFini', function (Blueprint $table) {
             $table->unsignedBigInteger('Id_Utilisateur');
             $table->unsignedBigInteger('Id_Manga')->index('Id_Manga');
-            $table->integer('note')->nullable();
+            $table->float('note')->nullable();
             $table->primary(['Id_Utilisateur', 'Id_Manga']);
         });
     }
